+++ score_keeper_server/modals/modal.js
@@ -15,27 +15,3 @@ const gameDataSchema = new mongoose.Schema({
 const GameData = mongoose.model('GameData', gameDataSchema);
 
 module.exports = GameData
-
-const newsSchema = new mongoose.Schema({
-    status: String,
-    totalResults: String,
-    articles: [
-        {
-        source: {
-            id: String,
-            name: String
-        },
-        author: String,
-        title: String,
-        description: String,
-        url: String,
-        urlToImage: String,
-        publishedAt: String,
-        content: String
-        }
-    ]
-})
-
-const News = mongoose.model('News', newsSchema);
-
-module.exports = News
